# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FlexBank Developer API (Phase 2)",
    "description": "Unified financial infrastructure platform providing identity, organizations, project environments, and secure API keys management.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/accounts/{accountId}/ledger": {
      "get": {
        "summary": "Fetch account ledger history",
        "description": "Lists cursor-paginated ledger entries for the specified account in reverse chronological order.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ledger history retrieved successfully"
          },
          "404": {
            "description": "Account not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}